refactor: :recycle: refactor requests page code in order to make the code cleaner
