+++ src/api/interfaces/index.d.ts
@@ -1,0 +1,1 @@
+export * from './contingency.interfaces';
