+++ src/pages/request/Request.tsx
@@ -1,121 +1,40 @@
-import { Badge, Button, notification, Table, Tabs } from 'antd';
-import { useEffect, useState } from 'react';
-import ApiHR from '../../api/ApiHR';
-import { useHandleError } from '../../hooks/useHandleError';
-import { columnsContigencyRequestInfo } from './table-design/contingency-request';
-import { ContingenciesTmHttp, ContingencyHttp } from '../../api/interfaces';
+import { Badge, Table, Tabs } from 'antd';
+import {
+  firstRow,
+  columnsContigencyRequestInfo,
+  lastRow,
+} from './table-design';
 import { ModalReject } from '../../components/modals/ModalReject';
 import { Loader } from '../../components';
-import { NotificationPlacement } from 'antd/es/notification/interface';
-import { CheckOutlined } from '@ant-design/icons';
 import { ModalAprove } from '../../components/modals/ModalAprove';
 import { ModalInfo } from '../../components/modals/ModalInfo';
-import Link from 'antd/es/typography/Link';
+import { useRequestPage } from '../../hooks';
 
 export const Request = () => {
-  //table variables
-  const [total, setTotal] = useState(0);
-  const [isLoading, setIsLoading] = useState(false); // show load table
-  const [contingencyRows, setContingencyRows] = useState<ContingencyHttp[]>([]);
-
-  //modal variables
-  const [isShowing, setIsShowing] = useState(false); // show reject or aprove loader
-  const [modalReject, setModalReject] = useState(false);
-  const [modalAprove, setModalAprove] = useState(false);
-  const [modalInfo, setModalInfo] = useState(false);
-  //data to set folio and id to aprove or reject
-  const [contingency, setContingency] = useState<ContingencyHttp>({
-    _id: '',
-    folio: '',
-    id_employee: 0,
-    name_employee: '',
-    date: '',
-    half_day: false,
-    status: '',
-    comments: '',
-    observations: '',
-    id_tm: 0,
-    createdAt: '',
-    updatedAt: '',
-    __v: 0,
-  });
-  //TODO: change contextHolder
-  //notifications
-  const { contextHolder: errorNotificacion, setServerError } = useHandleError();
-  const [api, contextHolder] = notification.useNotification();
-
-  const openNotification = (
-    placement: NotificationPlacement,
-    messages: string | string[],
-  ) => {
-    api.success({
-      message: 'Successful Operation',
-      description: messages,
-      placement,
-    });
-  };
-
-  const getContingenciesByPage = async (page?: number) => {
-    try {
-      setIsLoading(true);
-      const { data } = await ApiHR<ContingenciesTmHttp>(
-        `/contingencies-tm/requests?page=${page ?? 1}`,
-      );
-      setContingencyRows(data.docs);
-      setTotal(data.totalDocs);
-      setIsLoading(false);
-    } catch (error: any) {
-      setIsLoading(false);
-      setServerError(error);
-    }
-  };
-
-  const updateContingencyStatus = async (data?: { observations: string }) => {
-    try {
-      const query = data
-        ? { status: 'rejected', ...data }
-        : { status: 'approved' };
-      setIsShowing(true);
-      await ApiHR.patch(
-        `/contingencies-tm/requests/update-status/${contingency._id}`,
-        query,
-      );
-      setIsShowing(false); //show the loader
-      getContingenciesByPage(); // refresh the table
-      // close modal and show notification
-      if (data) {
-        setModalReject(false);
-        openNotification('top', 'Contingency rejected');
-      } else {
-        setModalAprove(false);
-        openNotification('top', 'Contingency approved');
-      }
-    } catch (error: any) {
-      setIsShowing(false);
-      setServerError(error);
-    }
-  };
-
-  const setParams = async ({
-    record,
-    openModal,
-  }: {
-    record: ContingencyHttp;
-    openModal: (param: boolean) => void;
-  }) => {
-    setContingency(record);
-    openModal(true);
-  };
-
-  useEffect(() => {
-    getContingenciesByPage();
-  }, []);
+  //get variables from customhook
+  const {
+    contingency,
+    total,
+    isLoadingTable,
+    contingencyRows,
+    isLoadingRequest,
+    modalReject,
+    modalAprove,
+    modalInfo,
+    setModalInfo,
+    setModalAprove,
+    setModalReject,
+    getContingenciesByPage,
+    updateContingencyStatus,
+    setParams,
+    contextHolder,
+  } = useRequestPage();
 
   return (
     <>
       <h1>Pending Request</h1>
-      <Loader show={isShowing} />
-      {errorNotificacion}
+      {/* show loader outside table */}
+      <Loader show={isLoadingRequest} />
       {contextHolder}
       <Tabs
         defaultActiveKey="1"
@@ -134,7 +53,7 @@ export const Request = () => {
           {
             label: (
               <>
-                {/* set the comment in this way to avoid console warning */}
+                {/* TODO: warning is still appearing */}
                 <Badge status="warning" count={total}>
                   <div style={{ padding: '5px' }}>Contingency</div>
                 </Badge>
@@ -142,75 +61,32 @@ export const Request = () => {
             ),
             key: '2',
             children: (
-              <>
-                {contextHolder}
-                <Table
-                  loading={isLoading}
-                  columns={[
-                    {
-                      title: 'Folio',
-                      dataIndex: 'folio',
-                      render: (_, record) => (
-                        <Link
-                          onClick={() =>
-                            setParams({ record, openModal: setModalInfo })
-                          }
-                        >
-                          {record.folio}
-                        </Link>
-                      ),
-                    },
-                    ...columnsContigencyRequestInfo,
-                    {
-                      title: 'Actions',
-                      dataIndex: 'actions',
-                      render: (_, record) => (
-                        <>
-                          <Button
-                            type="primary"
-                            shape="circle"
-                            icon={<CheckOutlined />}
-                            style={{ background: 'green', margin: '2px' }}
-                            onClick={() =>
-                              setParams({
-                                record,
-                                openModal: setModalAprove,
-                              })
-                            }
-                          />
-                          <Button
-                            type="primary"
-                            shape="circle"
-                            icon={'X'}
-                            style={{ background: 'red', margin: '2px' }}
-                            onClick={() =>
-                              setParams({
-                                record,
-                                openModal: setModalReject,
-                              })
-                            }
-                          />
-                        </>
-                      ),
-                      align: 'center',
-                    },
-                  ]}
-                  rowKey={'_id'}
-                  dataSource={contingencyRows}
-                  onChange={() => getContingenciesByPage()}
-                  pagination={{
-                    pageSize: 5,
-                    total,
-                    onChange(page) {
-                      getContingenciesByPage(page);
-                    },
-                    hideOnSinglePage: true,
-                    showTotal(total, range) {
-                      return `${range[0]}-${range[1]} of ${total} items`;
-                    },
-                  }}
-                />
-              </>
+              <Table
+                loading={isLoadingTable}
+                columns={[
+                  ...firstRow({ setParams, setModalInfo }),
+                  ...columnsContigencyRequestInfo,
+                  ...lastRow({
+                    setParams,
+                    setModalAprove,
+                    setModalReject,
+                  }),
+                ]}
+                rowKey={'_id'}
+                dataSource={contingencyRows}
+                onChange={() => getContingenciesByPage()}
+                pagination={{
+                  pageSize: 5,
+                  total,
+                  onChange(page) {
+                    getContingenciesByPage(page);
+                  },
+                  hideOnSinglePage: true,
+                  showTotal(total, range) {
+                    return `${range[0]}-${range[1]} of ${total} items`;
+                  },
+                }}
+              />
             ),
           },
           {
@@ -225,7 +101,7 @@ export const Request = () => {
         ]}
       />
 
-      {/* Modals */}
+      {/*--------------------------------------- Modals ---------------------------------*/}
       <ModalAprove
         changeStatus={updateContingencyStatus}
         folio={contingency.folio}
