+++ src/pages/request/table-design/index.ts
@@ -1,0 +1,3 @@
+export { firstRow } from './first-row';
+export { lastRow } from './last-row';
+export { columnsContigencyRequestInfo } from './other-rows';
